Bot receives reply from user
After messaging a user with the role configured, the bot will wait a set amount
of time to receive a reply from the user. The bot will only except a message if
it is sent in DMs (The channel and messaging of the user will later be
configurable). Upon reveiving a reply from the user, the bot will then echos the
message back to the user in the DM.