Message users at scheduled time and store response
Bot will message users, with specified role at
requested time from config, and store the users
reply in the database, if a reply is expected.

If a reponse from the user is not expected (i.e.
there is already a response stored in the database
for the most recent question asked by the bot),
then the bot ignores the message.