Ask user multiple questions and store answers
Allow the bot to ask the next question after the
user has replied to the current question. All
the messages, which are sent directly after a
question, are stored in the database.

This means that messages being sent after
all questions have been answered are not stored
in the database.